fix: renaming the action and adding docs
