--- aep-sync/.lintstagedrc
@@ -1,3 +1,0 @@
-{
-  "*.{js,jsx,css,scss,ts,tsx,json,yml}": "prettier --write"
-}
