--- aep-sync/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "aem-experiments-gh-action-aep-sync",
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/index.js",
-  "directories": {
-    "lib": "lib"
-  },
-  "files": [
-    "dist/index.js",
-    "dist/licenses.txt",
-    "lib"
-  ],
-  "scripts": {
-    "build": "ncc build lib/*.js --license licenses.txt",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@actions/core": "1.10.1",
-    "@vercel/ncc": "0.38.1"
-  },
-  "dependencies": {
-    "@actions/github": "6.0.0",
-    "@actions/http-client": "2.2.0",
-    "@adobe/aio-lib-ims": "6.5.0",
-    "js-base64": "3.7.5",
-    "jsdom": "22.1.0"
-  }
-}
