+++ package.json
@@ -16,5 +16,5 @@
     "nx": "17.1.2",
     "prettier": "3.1.0"
   },
-  "workspaces": ["aep-sync"]
+  "workspaces": ["segments-sync"]
 }
