+++ segments-sync/.lintstagedrc
@@ -1,0 +1,3 @@
+{
+  "*.{js,jsx,css,scss,ts,tsx,json,yml}": "prettier --write"
+}
