+++ segments-sync/action.yml
@@ -1,0 +1,43 @@
+name: 'AEM Experiment AEP Sync'
+description: 'Sync AEP experiments with AEM EDS'
+inputs:
+  aio_console_client_id:
+    description: The client id from the credentials for your project in the developer console
+    required: true
+  aio_console_client_secret:
+    description: The client secret from the credentials for your project in the developer console
+    required: true
+  aio_console_ims_org_id:
+    description: The ims org id from the credentials for your project in the developer console
+    required: true
+  aio_console_technical_account_email:
+    description: The technical account email from the credentials for your project in the developer console
+    required: true
+  aio_console_technical_account_id:
+    description: The technical account id from the credentials for your project in the developer console
+    required: true
+  git_pat_token:
+    description: The PAT token to be used to write the segments to the git repo
+    required: true
+  git_repo:
+    description: The owner/repo to push the segments to
+    required: true
+  git_ref:
+    description: The branch to push the segments to
+    required: false
+    default: 'main'
+  segments_path:
+    description: The path of the segments file in the git repo
+    required: false
+    default: 'segments.json'
+  segments_api_endpoint:
+    description: The endpoint to fetch the segment definitions from
+    required: false
+    default: 'https://platform.adobe.io/data/core/ups/segment/definitions'
+  sandbox_name:
+    description: The sandbox name to fetch the segments from
+    required: false
+    default: 'prod'
+runs:
+  using: 'node20'
+  main: 'dist/index.js'
